pr comments addressed
