--- packages/web-app/src/components/executionWidget/withdrawCard.tsx
@@ -1,53 +1,0 @@
-import React from 'react';
-import {CardTransfer, CardToken} from '@aragon/ui-components';
-import {useTranslation} from 'react-i18next';
-import styled from 'styled-components';
-
-import {ActionWithdraw} from 'utils/types';
-import {AccordionMethod} from 'components/accordionMethod';
-
-export const WithdrawCard: React.FC<{
-  action: ActionWithdraw;
-  daoName: string;
-}> = ({action, daoName}) => {
-  const {t} = useTranslation();
-
-  return (
-    <AccordionMethod
-      type="execution-widget"
-      methodName={t('AddActionModal.withdrawAssets')}
-      smartContractName={t('labels.aragonCore')}
-      verified
-      methodDescription={t('AddActionModal.withdrawAssetsActionSubtitle')}
-    >
-      <Container>
-        <CardTransfer
-          to={action.to}
-          from={daoName}
-          toLabel={t('labels.to')}
-          fromLabel={t('labels.from')}
-        />
-        <CardToken
-          tokenName={action.tokenName}
-          tokenImageUrl={action.tokenImgUrl}
-          tokenSymbol={action.tokenSymbol}
-          tokenCount={action.amount}
-          treasuryShare={
-            action.tokenPrice
-              ? new Intl.NumberFormat('en-US', {
-                  style: 'currency',
-                  currency: 'USD',
-                }).format(action.tokenPrice * action.amount)
-              : t('finance.unknownUSDValue')
-          }
-          type={'transfer'}
-        />
-      </Container>
-    </AccordionMethod>
-  );
-};
-
-const Container = styled.div.attrs({
-  className:
-    'bg-ui-50 rounded-b-xl border border-t-0 border-ui-100 space-y-3 p-3',
-})``;
